refactor: move poller.rs test module to poller_tests.rs (4702 LOC)
The 143-test module (78% of poller.rs) is extracted to a separate file
via #[path = "poller_tests.rs"]. The super::* import path is preserved
so no test code changes are needed — pure file split.

poller.rs: 5927 → 1226 lines (prod code only)
poller_tests.rs: 4702 lines (all 143 tests)


Agend-Agent: fixup-dev
Agend-Branch: refactor/cicd-medium-findings
Agend-Issued-At: 2026-05-23T17:19:52.645141+00:00
